content-service/admin/config/settings.py
import os
from pathlib import Path

from dotenv import load_dotenv
from split_settings.tools import include


load_dotenv()

# Build paths inside the project like this: BASE_DIR / 'subdir'.
BASE_DIR = Path(__file__).resolve().parent.parent

# Database
include(
    'components/database.py',
    'components/security.py',
    'components/app.py',
    'components/password.py',
    'components/internationalization.py',
)

# Static files (CSS, JavaScript, Images)
# https://docs.djangoproject.com/en/3.2/howto/static-files/

STATIC_URL = '/static/'

STATIC_ROOT = os.path.join(BASE_DIR, "static")

MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
MEDIA_URL = 'media/'
